requirements.txt
streamlit>=1.30
python-dotenv>=1.0
requests>=2.31
python-docx>=1.1
pypdf>=4.0
